HADOOP-1628 merged into branch-0.14.
git-svn-id: https://svn.apache.org/repos/asf/lucene/hadoop/branches/branch-0.14@559628 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/dfs/DFSClient.java
@@ -1665,7 +1665,7 @@ class DFSClient implements FSConstants {
           numSuccessfulWrites++;
 
           //We should wait for response from the receiver.
-          int reply = blockReplyStream.readByte();
+          int reply = blockReplyStream.readShort();
           if ( reply == OP_STATUS_SUCCESS ||
               ( reply == OP_STATUS_ERROR_EXISTS &&
                   numSuccessfulWrites > 1 ) ) {
